# scilvera_implementation/scipubsi/output/User/openapi.json
{"openapi": "3.0.0", "info": {"version": "0.0.1b", "title": "User", "description": ""}, "servers": [{"url": "http://localhost:9095/api/u"}], "paths": {"/user": {"post": {"description": "Creates a new user.", "operationId": "createUser", "requestBody": {"description": "User to add.", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/User"}}}}, "responses": {"200": {"$ref": "#components/responses/UserOK"}}}}, "/user/id": {"put": {"description": "Update User object.", "operationId": "updateUser", "parameters": [{"name": "id", "in": "path", "description": "ID of user to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/UserBody"}, "responses": {"200": {"$ref": "#components/responses/UserOK"}}}, "get": {"description": "Read User object.", "operationId": "readUser", "parameters": [{"name": "id", "in": "path", "description": "ID of user to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/UserBody"}, "responses": {"200": {"$ref": "#components/responses/UserOK"}}}, "delete": {"description": "Delete User object.", "operationId": "deleteUser", "parameters": [{"name": "id", "in": "path", "description": "ID of user to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/UserBody"}, "responses": {"200": {"$ref": "#components/responses/UserOK"}}}}, "/userexists/{username}": {"get": {"description": "", "operationId": "userExists", "parameters": [{"name": "username", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "JSON representation of boolean object", "content": {"application/json": {"schema": {"type": "boolean"}}}}}}}}, "components": {"schemas": {"User": {"type": "object", "properties": {"username": {"type": "string"}, "password": {"type": "string"}, "email": {"type": "string"}, "firstname": {"type": "string"}, "lastname": {"type": "string"}}, "required": ["username", "password", "firstname", "lastname"]}}, "requestBodies": {"UserBody": {"description": "JSON representation of User object", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/User"}}}}}, "responses": {"UserOK": {"description": "JSON representation of User object", "content": {"application/json": {"schema": {"$ref": "#components/schemas/User"}}}}}}}